requirements.txt
aiogram>=3.4.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
requests>=2.31.0
dateparser>=1.2.0
